[fix] ENZEL alignment tab should pause stream when exiting the tab
It would pause the streams when changing mode inside the tab... but not
pause them when going to another tab, like the "localization" tab.
That's very confusing, because this leaves the detectors active, which
can cause lots of odd behaviour in the other procedures.

=> rename pauseAllStreams() to pauseStreams() of the stream bar
controller so that it has the same API as the other stream bar controllers.