[perf] tileacq: don't switch external mode between each tile
If there is an automatic external mode (on the e-beam), then during the
acquisition, it will "flicker" at each tile. This can cause extra delay,
and also some artifacts. So force it on at the beginning of the whole
acquisition, and only reset it after the entire series is complete.