[fix] Avoid infinite loop when view mpp is changed
In some very rare cases, a view resize would cause a change in the mpp
(aka zoom level), which would update the FoV, and cause again a change
in the mpp... infinitely. This caused a hand of the GUI.

In particular this happened *sometimes* when starting the SECOM GUI.
However, it could have happened on almost any system.

=> Have a single function that listen to the mpp change, not two!