[fix] Make the beamshift correction save path unique (#2950)
With the volume EM changes the ROA name does not have to be unique anymore. Since the beamshift correction save path was only based on the ROA name, this caused all the images of a project to be saved in the same directory, and fields with the same id got overwritten.
This fix extends the save path to include the project name, the roa name and the slice idx.