[fix] SPARC acq: handle floating point error in comparison
In some very rare cases, comparing the dwell time of the emitter with
the dwell time of the stream wouldn't perfectly match, with confusing
error messages like this:
Traceback (most recent call last):
  File "/home/sparc/development/odemis/src/odemis/acq/stream/_sync.py", line 2433, in _runAcquisition
    raise IOError("Expected hw dt = %f but got %f" % (px_time, self._emitter.dwellTime.value))
OSError: Expected hw dt = 0.000100 but got 0.000100

That was due to floating point error. => Change to an almost equal test.