[fix] limit how quickly turnOnLight() acquires images
turnOnLight() would continuously acquire images until it detects light.
This is a little over the top, as it can take several seconds and it's
really not in a rush. If the CCD is set to short exposure times, that
could lead to 100 fps, which causes high CPU usage, with no advantage.

=> wait at least 100ms per frame, so at most it will be 10fps.

As a side effect, that will help some test cases which were setting
the exposure time to the minimum, and which relied on andorcam2. That
driver seems to have a memory leak when doing many DataFlow.get() in a
row.