[feat] [METEOR-590] Add TFS2 posture manager with robust feature functionality (#2730)
Improved stage movement on the TFS SEM for the METEOR. 

Read raw z coordinate system in TFS systems and image on fixed plane when performing FM imaging. The main advantage is that when moving back and forth between FM and SEM, the stage should always be at the same Z, both in the SEM posture, and the FM posture. This ensures the FoV always corresponds to the same point in FM and SEM.

For the long term, this should also provide a better base to support imaging in other postures than FM.