[fix] gui SPARC: time estimation should be updated when drift correction is adjusted
When the drift correction is activated (ie, .roa goes from UNDEFINED_ROI to
another value), or the period is changed, the total acquisition time is
affected (a little bit). So need to automatically update it, like any
other VA affecting the acquisition.

Note that the dwell time is linked to the SEM survey dwell time. So in
theory, it's not really necessary to also listen to it, as the SEM dwell
time would already trigger the time estimation. However that's not a big
deal, and ensure that even if we change them to not be linked, the time
will still be correctly estimated.