[feat] TemporalSpectrum live acquisition protection
Every time an streakcam image is received, check that the intensity
level is not too high. If too many pixels are too high, with a threshold
defined in the microscope file, then the acquisition is stopped (and
the data acquired so far is saved).

The maximum safe intensity level can be defined in the microscope file,
by adding a metadata on the streak-ccd component like this:
metadata: {
    CALIB: {"intensity_limit": 50000},  # counts/s
},