[fix] driver pwrmccdaq: simulate better connection of DIO ports A <-> B
The simulator simulates port A being connected to port B. So writting to
a bit on port A causes the equivalent bit on port B to read the same
value (and inversely too).

This worked, but it broke the possibility to read what value is
currently set. => write both to port A & B.

Also *document* this behaviour.

Also, as the test cases relied on such behaviour on the real hardware (by
physically connecting some pins together). The tests also work on the
simulator. So no need to disable it... However that test was broken, so
fix it too.