[refactor] remove old scripts not useful anymore
errorcodegen: was used to generate some python code of the andorcam2
driver.
dmeepromconfig: potential script to write to the EEPROM of the SPARC
modules. Eventually the SPARC modules were slimmed down and don't have
EEPROM. Never used in production.
yaml-name-update: use to help the update of some microscope files, in
2015.