[feat] acq SEMCCDMDStream use startScan Event to synchronize CCD with e-beam
From the momement that the acquisition ask the e-beam scanner to start
and the e-beam is actually at the right position, it takes *some* time.
It's in the order of a few ms. So far, the only e-beam scanner was
semcomedi, and it was (surprisingly) constant in terms of time. It would
take between 2 and 5 ms. So it was good enough to just always wait 5ms.

We now also have semnidaq, and the NI-DAQ library introduces a lot longer latency.
While it can be as "fast" as 10ms, it sometimes takes 40ms to start a
scan. There are probably ways to make it a little faster, but it's
unlikely it will be possible to always keep all scanners < 5ms.

=> Use a new Event, startScan, for the e-beam scanner to report the
e-beam is ready. Then, we can use it to directly start a CCD
acquisition.