[feat] SettingsObserver should also store axes "user-friendly" info if available
For some axes, the "position" is an abritrary number. For instance, on
the spectrograph, the band is 1->4. This is not helpful for the user.
However, the .axes contains the information to map that value (eg, 1) to
a user-friendly description (eg, "1200 l/mm (blaze: 500.0)").

Several SPARC users requested for this info to be stored too.

=> When available, also store the axis definition.