[fix] gui HW_SETTING_CONFIG too specific for e-beam settings
Commit e5fc2889 (minor gui/model changes for fibsem tab) changed the
e-beam settings to very specific. In particular, it forced the
probeCurrent VA to *always* be a slider. However, on some hardware this
VA can be a set of choices. This caused the GUI from failing to display
anything at all in such case.

When no control_type is specified, the type is automatically selected
using determine_default_control(). It should work for both cases of
Continuous and Enumerated VAs.