[fix] driver ephemeron: work with the standard version of asyncua package
Until now, the simulator depended on a "fix" provided by Ephemeron for
the asyncua package. However, it's not clear whether that fix is ever
going to be integrated into the official version. That makes the whole
environment really difficult to set-up.

=> Carry the "fix" directly in the code, by overriding the standard
StateMachine class.

Also add some information about the dependency on asyncua.