[feat] SEMMDStream: handle "independent" detector streams which adjust the scanner dwell time
The IndependentEBICStream picks a dwell time for the scanner a little
larger than the main "dwellTime", which represents the detector dwell
time. On streams with 3 detectors (aka "folded streams"), this would
either confuse the acquisition code (because the dwell time is different
than the one expected) or would not work (because the other stream would
override the value).
=> explicitly detect that the detector is "independent" and apply its
settings last, to make sure they are used in priority.