[fix] SPARC acquisition with integration time could estimate time incorrectly
When there is an integration time, the actual time is exposure time *
integration count.

At the moment the hardware parameters are set, it's fine to use the
actual CCD exposure time to compute the integration time. However, when
estimating the acquisition time, the CCD VA can be pretty much any
value, if the stream is not playing. So we shouldn't use that VA.
Instead just trust the pre-computed integrationCounts.

=> Avoids odd warnings