[feat] GUI SPARC: at init also blank the beam if the interlock is triggered at start
We try to not cause changes in the hardware when the GUI starts, but we
make an exception here. This way, if the interlock is triggered, the
user will immediately notice, via a notification message. This can also
blank the e-beam, which maybe the user doesn't expect... but it's also
safer this way, as it's a reminder to the user that they shouldn't have
the cover opened while the e-beam is active.