[doc] SPARC GUI: update streakcam calibration procedure description
Based on more experience, and related to the new streakcam, the
calibration procedure has been updated by Herman.
It should be fine to not keep the old version (for the old hardware), as
the old hardware is used only on one install, where users are very
familiar with the procedure. Moreover, the new procedure is still
correct, just some steps have been removed (because with the new
hardware they happen automatically, or are not needed).