[feat] add transform support for METEOR Tescan1
Requires support for shear and scale, so introduce this too.
It seems the 3d transformations were using the inverted calculation
"p @ trans_mat" instead of the standard "trans_mat @ p". With the
rotation, that worked by using the inverse of the matrix, so not really
noticeable. However, with the shear and scale, we need the right order
to have it work (the standard way).