[feat] New plugin spectrum_laser_protection, for laser protection with SPARC FPLM
Changes the behaviour of the GUI, so that on a SPARC, it's not possible to turn on the laser
of a spectrum stream when the parabolic mirror is not engaged. The idea is that, although unlikely,
the user could go to the "acquisition" tab without having engaged the mirror (typically, it's not useful,
but the GUI allows this). In this case, the input laser (coming from via a FPLM module) could enter
the chamber and hit and damage some of the detectors.
The way it works in practice is that whenever the acquisition tab is shown, it checks the parabolic mirror
position. If it's not engaged it:
* sets the spectrum stream light power to 0, and disable the entry. So it's impossible to turn it on.
* when a new spectrum stream is added, it disables the power entry.
If the mirror is engaged, the power entry of all spectrum streams is re-enabled.