[feat] Don't show read-only VAs in the HW settings by default
Read-only settings are typically not useful to the user (as they cannot
be changed). So by default there is not need to show them in the
hardware settings panel. It's still possible to force showing them by
explicitly listing them in the HW_SETTINGS_CONFIG.

In practice, that shouldn't change any of the current system behaviour,
as for any VA we care, we always list them in the HW_SETTINGS_CONFIG, to
ensure it's displayed in the right order.
However, for the ebeam-gun-exciter, it's useful, as we explicitly added
read-only VA to store extra information in the metadata.