[fix] MeteorTescan1PostureManager could go directly from UNKNOWN to FM_IMAGING
When requesting a move from UNKNOWN to FM_IMAGING, the code *tried* to first go
to SEM posture, but a typo made it actually go directly to FM_IMAGING.
=> fix the variable name to do the right thing.

Note that currently the GUI actually doesn't allow going from UNKNOW to FM_IMAGING,
so that not a big deal. Still, if the code is there, it should do the
right thing.