[fix] dataio.tiff: loading file with mix of z-stack and non z-stack fail
When loading a z-stack, the center position in Z must be computed and
updated to all image *of the z-stack*. It used to update to all images
in the TIFF file, which caused various issues. One of the main one being
that if another image was a preview image, it would not have metadata,
and the whole process would fail.
It could also potential put incorrect metadata.

Fixes issues like this:
2025-05-19 16:42:49,661 ERROR   log:41: Failed to decode OME XML
Traceback (most recent call last):
  File "/home/piel/development/odemis/src/odemis/dataio/tiff.py", line 2596, in _getAllOMEDataArrayShadows
    _updateMDFromOME(omeroot, data)
  File "/home/piel/development/odemis/src/odemis/dataio/tiff.py", line 993, in _updateMDFromOME
    pos = da.metadata[model.MD_POS]
AttributeError: 'NoneType' object has no attribute 'metadata'