[fix] sparc2 autofocus test cases
TestSparc2AutoFocus_2.test_cancel and TestSparc2AutoFocus_2::test_one_det failed because the stage tried to move out of range.
This happened because at initialization of the test class the good focus is set based on the current stage position. In the test the focus moves 200um away from the good focus.
Because the backend does not restart when running multiple test classes with the same backend, the second class that was run had the wrong good focus position.