[fix] Optical Path Manager test for FSLM had wrong expectation about spec-switch
Commit 8ca88e85bd (FSLM simulator had incorrect affects) fixed the
affects of the FSLM simulator. This meants the optical path manager can
now move the axes as expected.

However, the test case for the light-in-align was expecting that the
spec-switch wasn't moved. That's incorrect. That axis should
automatically be adjusted so that it's possible to observe the signal on
the internal CCD.
=> Change the test case to verify that the axis position is always
reset.