[fix] camera testing: add a bit of time after each test
Some test cases would start with changing the settings and do data acquisition.
However, on some hardware, like the simstreakcam, if one stop/start very
briefly the acquisition, it can happen that the acquisition is not
stopped at all. From a driver's perspective, it's an acceptable
behaviour. But that meant some test cases, on some hardware, sometimes
failed due to receiving data with the wrong shape or metadata.

=> wait a little bit after the end of the test case to be certain the
acquisition is stopped (and therefore the next data acquired will be
with the new settings).
In particular, this fixes failures sometimes happening in
TestSimStreakCamGenericCamSynchronized.test_software_trigger .