[fix] SPARC streak cam alignment should start with the slit almost closed
It's safer for the camera to let the spectrograph slit "almost" closed (ie, not fully
opened via the slit-in-big). The user can still change it from the GUI,
so this is not a big change.

Note that this affects only the SPARCs where there streakcam is after the
standard spectrograph, not the ones which uses the light tunnel.