[fix] GUI: on wxPython 4.1+, don't show warning box every time a dialog box is opened
With the latest version of wxPython, every time the user would open a
dialog box containing some number control, there would be first show a
warning box stating: "Could not transfer data to window".

It turns out it's coming from a new feature of the Validators which
allow to copy the data. The base class is very pessimistic and assumes
it failed. So every inheriting class must override it and return True to
indicate there was no data to copy anyway.