[fix] TFS adapter drivers failed to transfer new version
Commit 76ded477 (add autoscript client for openfibsem) adjusted the code
to detect a new version of the xtadapter and transfer it, so that it
could be shared between the xt_client and autoscript_client drivers.

However, while the transfer_latest_package() was generalised, the caller
wasn't updated. This broke the transfer mechanism.

Just provide the same function on both clients, and use the same old
caller.