[fix] GUI analysis tab showed too many viewports on wxPython 4.1+
On wxPython 4.1+, the hidden flag in XRC has to be explicitly treated by
XmlRessourceHandlers. Commit d82ad870df (XmlResourceHandler now
explicitly handle the hidden flag) took care of it... except it missed
the MicroscopeViewport class.

This caused the analysis tab, which has 12 viewports, to show all of
them at start up, and when opening a new file.

=> Also treat the hidden flag in this handler.

Also detect such error (which could be also due to modifying the XRC
file) of more than 4 viewports at init, and warn about it. This should
make it easier to find out next time.