[fix] show acquired sem overview stream panel
When the sem overview streams are acquired in the fibsem tab, it is also added in the correlation tab and the localizaton tab. If the streams are updated in the correlation tab, they are also updated in the localization and the fibsem tabs. The update means addition of a new stream, deletion of the existing stream, modification in the correction metadata and lastly the reset of the correction metadata