[feat] add functionality to shift the phase of the descanner setpoints
To align the e-beam scanner movement with the descanner mirrors, add a shift that is calibrated in fastem-calibrations.
The phase shift differs per multibeam dwell time, add the calibrated values to the metadata to be able to update the phase shift when the multibeam dwell time is updated.

Update after PR comments

- added unit and explanation for descanner shift VA
- catch KeyError in phase shift is not defined in MD_CALIB

fixes after PR comments

- add validation check in updateMetadata for phase_shift_md
- update comments and docstrings
- use util function for finding closest dwell time