[fix] Update scripts and documentation for newer version of Conda on Windows
Now that Python 3.6 is not supported anymore, we really needed to bump
the Windows build to a newer version of Python, and conda.

Update the instructions use the latest version of Conda, with Python
3.10.

Also update the installation scripts for PyInstaller v6.0, and newer
version of setuptools (v80).
In particular, the installation files are stored in a subfolder
"_internal", so so code needs to be aware of this path. Also, to make
the transition smooth, we need to delete the files of the old Odemis
install, to be certain they are not loaded first.