[fix] odemis-select-channel: handle DEB822 repo format
Since Ubuntu 20.04, apt supports two types of format for the PPAs: the
"legacy" ones (one line) and the DEB822 ones on multiple lines.
Since Ubuntu 24.04, the default is to use DEB822. So we need to be able
to read and modify this type of file too.

Also fix a few other issues:
* When activating dev channel, make sure the DEPATH is uncommented.
* After the change is done (which can take up to a minute in the
  background), display a pop up window to indicate it's ready.
* Never call explicitly commands via sudo. The --select option has to be
  run with the right priviledges.