[ci] Add logging timeouts and other errors to runtests.sh
The status of calling the test case from bash was logged in the full test report, but nothing was done with it. Now if the status is not 0 (OK) it is also copied to the summary.
Especially for timeouts this is usefull to know, because if test cases time out failures can get hidden.