[feat] SPARC GUI: force the SEM survey stream to always start with the same settings
The original idea was to pick the settings from the current hardware
settings. So this would allow to adjust the default settings in the
microscope file, and if the GUI is closed and restarted the previous
settings values would be used.

Nice idea, but it doesn't work well in practice because after an
acquisition the settings are changed to "weird" values. So the user ends
up sometimes with weirds values, and in general with a little bit
"random" values.

Instead, always use scale 8 and dwell time 1 µs. That's roughly 0.5s
acquisition on most systems. That should be good
enough for most of the usage, and be a lot let surprising.