[fix] SPARC acquisition: make it possible to go from spot tool -> RoA
Until now, if a stream using the spot mode is playing (eg, spectrum),
and the user presses the "RoA" tool, then the stream pauses, the spot
tool is disabled... and the RoA tool is disabled too. Then the user has
to press a second time on the RoA tool.

That's because when stopping the spot stream, the scheduler would
incondionally reset the tool mode. Instead, it should just make sure
the tool mode is not "spot" any more.