[fix] semnidaq scan*TTL VAs were not properly applied
The code tries to convert a boolean to a (arbitrary) int. To do so, it
did "b & i", but that doesn't work as intended, as it converts first i
to a boolean. Instead, one should do "b * i".

This meant that on real hardware, the TTLs were not applied (on the
correct port).