[fix] METEOR: don't ask the user to update feature positions if in Odemis standard
For Odemis Advanced, a new feature was added to allow the user decide
whether to update the position in all posture or only the current one.
However, it's actually only useful if the position had been set explicitly in a different posture.

It's annoying to have this question every time you move a feature.
Especially, on Odemis standard, it's only possible to move the feature
in the FM posture, so the user always wants to say "yes".

Ideally, we would detect for each feature in how many posture was the
posture set. For now, we detect that if there are only 2 postures,
it's a standard Odemis, so it's only possible to set the position in FM,
so it's never needed to ask the user.