[feat] MeteorPostureManager SampleStage also translate the position
The goal is to have the same position, in absolute value, for the same
point on the sample, independent of the posture. So at init, the offset
needed for each posture is computed, and afterwards used to translate
the reported position.

To make the values a little bit more relatable to the SEM position, the
origin of the sample stage is selected so that the center of GRID 1 in
the SampleStage is at the same position as in the (SEM) stage bare
coordinate system.