[fix] spot-grid freezes when there was no grid visible in the live stream
The spot-grid script contained a try except, but only had an exception for ValueErrors. In odemis.util.spot._get_subimage an IndexError gets raised when the position of the spot is refined and that is too close to the edge of the image cropped around the spot.
This happened when there was no pattern visible on the stream but only noise. This fix also adds an exception for IndexErrors and a general Exceptions.
This means the spot grid will not get drawn when there are no spots and the live stream will also not freeze.

IndexError Position too close to the edge of the image.