[test] fix n_jobs usage in spot_test for ubuntu 24
In newer versions of scipy kdtree the usage of the `n_jobs` argument is replaced by the `workers` argument. This was already fixed in the code, but not yet in the test case.

This fixes the following error:
```
spot_test.py:335: in test_multiple
    distances, indices = tree.query(ji, k=1, n_jobs=-1)
_ckdtree.pyx:786: in scipy.spatial._ckdtree.cKDTree.query
    ???
_ckdtree.pyx:388: in scipy.spatial._ckdtree.get_num_workers
    ???
E   TypeError: Unexpected keyword argument {'n_jobs': -1}
```