[feat] util img: dedicated function for converting RGB data to RGB8
Converting from RGB to RGB should be simple... and it used to be handled
by essentially copying the data. Just a basic code to handle tint.
However, this works well only if the data is already uint8.
For uint16, this didn't work. Also, the range wasn't computed.

=> introduce projectYXC2RGB8() to handle explicitly such method, and use
in when the data is in "YXC" format.
In particular, this helps opening Tescan RGB 16-bits images.