[test] Meteor PostureManager: don't try to switch from loading -> Grid
Switching between grid should only be possible in posture where the
sample stage can move along the sample plane. So no in LOADING posture.

Until now, on some systems, it was possible to go directly from LOADING
to a Grid, in which case, it would go to SEM posture, at the given GRID.
That's "easy" to support because switching between LOADING and SEM actually goes to
GRID 1, however, it breaks the idea that switching between grid is
orthogonal to switching postures. In particular, now the Tescan raise an
exception because it's not a supported switch.

=> Drop related tests