[feat] METEOR GUI: automatically delete empty project folder
When closing the GUI, if the project folder, it means the user didn't
use it. No need to keep it. This avoids filling up the main folder with
empty project folders, and allows to have automatically the same name as
the previous project when restarting the GUI, if it wasn't used.