[fix] GUI FIBSEM tab: ensure the posture shown is always the current one
The function is called whenever the stage-bare position changes.
It used to read the current posture based on the .current_posture VA.
However, .current_posture VA is also updated via callback on a
stage-bare position. So there was 50% chance it's updated after this
function. In such case, the function would read the posture as of the
previous position... which in some cases could be still a different
posture (typically, UNKNOWN).