[feat] GUI: open files using case insensitive wildcard
On Linux, the filenames are case sensitive. That also means the
files extensions are case sensitive. Until now, Odemis would open *.tif
and *.hdf5. So a file name blabla.TIF would not even be listed to the
user when opening a file.

=> Add the upper case version of every extension to support it.

Only show it when opening a file. When saving a file, it's fine to use
always lower case (which is the standard on Linux).

Note that this is not exactly case-insensitive, as extension like "Tif"
would still not be detected, but that's very unlikely to exist.