[fix] METEOR GUI: simplify warning message when switching posture
The GUI used to check that the current tilt is close to the expected
tilt. However, this check is already done by the PostureManager, and
will concider any large variation in tilt as outside of the standard
posture, which will report the posture as UNKNOWN, and not allow to
switch posture at all using the GUI.

So no need for the check anymore. It actually was triggering sometimes
erroneous messages just because it was too picky about the tilt angle
(as the hardware precision is a lot lower than just floating point
error).