[fix] Fix TOA hfw selection and overlap
For TOAs it was not possible to select a horizontal field width lower than 50 µm. And the overlap at that HFW was >50%
The TOA code internally calls the overview image code, the overlap in the overview is set to stage_precision / horzontalFoV.
But the stage_precision is hard coded to 29 µm, for a HFW of 50µm this is an overlap of 58%.
When the overview code was written it made sense to base the overlap on the stage precision, however at 2mm field of view there are more factors,
like beam deflection at the edges, influencing what overlap make sense. For now just hardcode the overlap.

FVEM-254