[fix] Represent invalid log range as cubic
During testing on the autoscript simulator, the probe current VA (represented by a log slider)
was found to have a minimum value of 0. This is problematic since log(0) is undefined.

To fix this in the general sense, the log representation now internally checks if there is
an invalid range, and if so, automatically switches to a cubic scale.