[feat] OPM: support time-correlator after the spectrograph
If a (optical fiber coupled to a) time-correlator is connected to the
spectrograph output, then the position of the spectograph input slit
matters.
Note: this actuator position will only be changed if it "affects"
(directly or indirectly) the time-correlator. So, on standard SPARCs
where the time-correlator is coupled before the spectrograph, this
change has no effect.